% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd_dissimilarity.R
\name{plot.intrait_tpd_dissim}
\alias{plot.intrait_tpd_dissim}
\title{Plot the TPD dissimilarity matrix as a heat map}
\usage{
\method{plot}{intrait_tpd_dissim}(x, col = grDevices::hcl.colors(50, "TealGrn", rev = TRUE), ...)
}
\arguments{
\item{x}{An object of class \code{"intrait_tpd_dissim"}.}

\item{col}{A vector of colours for increasing dissimilarity. Defaults to a
white-to-teal ramp.}

\item{...}{Passed to \code{\link[graphics:image]{graphics::image()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot the TPD dissimilarity matrix as a heat map
}
