% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_disparity.R
\name{print.intrait_disparity}
\alias{print.intrait_disparity}
\title{Print an \code{"intrait_disparity"} object}
\usage{
\method{print}{intrait_disparity}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"intrait_disparity"}, as returned by
\code{\link[=trait_disparity]{trait_disparity()}}.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an \code{"intrait_disparity"} object
}
