% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{print.intrait_measurement_error}
\alias{print.intrait_measurement_error}
\title{Print an \code{"intrait_measurement_error"} object}
\usage{
\method{print}{intrait_measurement_error}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"intrait_measurement_error"}, as returned by
\code{\link[=measurement_error]{measurement_error()}}.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an \code{"intrait_measurement_error"} object
}
