% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{load_fishmorph_phylogeny}
\alias{load_fishmorph_phylo_axes}
\alias{phylo_pcoa}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Rfishmorph}{\code{\link[Rfishmorph:load_fishmorph_phylo_axes]{load_fishmorph_phylo_axes()}}, \code{\link[Rfishmorph:load_fishmorph_phylogeny]{load_fishmorph_phylogeny()}}, \code{\link[Rfishmorph:phylo_pcoa]{phylo_pcoa()}}}
}}

