% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itv_index.R
\name{plot.intrait_itv}
\alias{plot.intrait_itv}
\title{Plot the interspecific/intraspecific variance breakdown}
\usage{
\method{plot}{intrait_itv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"intrait_itv"}, from \code{\link[=itv_index]{itv_index()}}.}

\item{...}{Further arguments passed to \code{\link[graphics:barplot]{graphics::barplot()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot the interspecific/intraspecific variance breakdown
}
