% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functional_space.R
\name{print.intrait_bootstrap_fspace}
\alias{print.intrait_bootstrap_fspace}
\alias{plot.intrait_bootstrap_fspace}
\title{Print and plot an \code{"intrait_bootstrap_fspace"} object}
\usage{
\method{print}{intrait_bootstrap_fspace}(x, ...)

\method{plot}{intrait_bootstrap_fspace}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"intrait_bootstrap_fspace"}, as returned by
\code{\link[=bootstrap_functional_space]{bootstrap_functional_space()}}.}

\item{...}{For \code{plot()}, further arguments passed to \code{\link[graphics:hist]{graphics::hist()}};
currently unused by \code{print()}.}
}
\value{
Invisibly returns \code{x}.

Invisibly returns \code{x}.
}
\description{
Print and plot an \code{"intrait_bootstrap_fspace"} object
}
